test: Move test_runner_close to own module
